# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RankYatra API - Competitive Exam Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /exams/{examId}/questions:
    get:
      operationId: getExamQuestions
      tags: [questions]
      summary: Get questions for a live exam (registered users only)
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Questions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
        "403":
          description: Not registered or exam not live
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: addQuestion
      tags: [questions]
      summary: Add a question to an exam (admin only)
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionBody"
      responses:
        "201":
          description: Question added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /exams/{examId}/questions/{questionId}:
    patch:
      operationId: updateQuestion
      tags: [questions]
      summary: Update a question (admin only)
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
        - name: questionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionBody"
      responses:
        "200":
          description: Updated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"

    delete:
      operationId: deleteQuestion
      tags: [questions]
      summary: Delete a question (admin only)
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
        - name: questionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── REGISTRATIONS ───────────────────────────────────────────────────────────
...
  /exams/{examId}/submit:
    post:
      operationId: submitExam
      tags: [submissions]
      summary: Submit answers for an exam
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitExamBody"
      responses:
        "200":
          description: Submission result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResult"
        "403":
          description: Not registered or exam not active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /exams/{examId}/my-result:
    get:
      operationId: getMyResult
      tags: [submissions]
      summary: Get current user's result for an exam
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResult"

  # ─── LEADERBOARD ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    # ─── AUTH ─────────────────────────────────────────────────────────────────
...
    Question:
      type: object
      properties:
        id:
          type: integer
        examId:
          type: integer
        questionText:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctOption:
          type: string
          enum: [A, B, C, D]
        orderIndex:
          type: integer
      required: [id, examId, questionText, optionA, optionB, optionC, optionD, correctOption, orderIndex]
    CreateQuestionBody:
      type: object
      properties:
        questionText:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctOption:
          type: string
          enum: [A, B, C, D]
        orderIndex:
          type: integer
      required: [questionText, optionA, optionB, optionC, optionD, correctOption, orderIndex]

    # ─── REGISTRATION ────────────────────────────────────────────────────────
...
    SubmitExamBody:
      type: object
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/Answer"
        timeTakenSeconds:
          type: integer
      required: [answers, timeTakenSeconds]
    Answer:
      type: object
      properties:
        questionId:
          type: integer
        selectedOption:
          type: string
          enum: [A, B, C, D]
      required: [questionId, selectedOption]
    SubmissionResult:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        examId:
          type: integer
        score:
          type: integer
        totalQuestions:
          type: integer
        correctAnswers:
          type: integer
        timeTakenSeconds:
          type: integer
        rank:
          type: ["integer", "null"]
        submittedAt:
          type: string
      required: [id, userId, examId, score, totalQuestions, correctAnswers, timeTakenSeconds, submittedAt]

    # ─── LEADERBOARD ─────────────────────────────────────────────────────────
...